+++ example/lib/main.dart
@@ -2,7 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:mobile_scanner_example/barcode_list_scanner_controller.dart';
 import 'package:mobile_scanner_example/barcode_scanner_controller.dart';
 import 'package:mobile_scanner_example/barcode_scanner_returning_image.dart';
+import 'package:mobile_scanner_example/barcode_scanner_window.dart';
 import 'package:mobile_scanner_example/barcode_scanner_without_controller.dart';
+import 'package:mobile_scanner_example/barcode_scanner_zoom.dart';
 
 void main() => runApp(const MaterialApp(home: MyHome()));
 
@@ -44,6 +46,16 @@ class MyHome extends StatelessWidget {
               onPressed: () {
                 Navigator.of(context).push(
                   MaterialPageRoute(
+                    builder: (context) => const BarcodeScannerWithScanWindow(),
+                  ),
+                );
+              },
+              child: const Text('MobileScanner with ScanWindow'),
+            ),
+            ElevatedButton(
+              onPressed: () {
+                Navigator.of(context).push(
+                  MaterialPageRoute(
                     builder: (context) => const BarcodeScannerReturningImage(),
                   ),
                 );
@@ -61,6 +73,16 @@ class MyHome extends StatelessWidget {
                 );
               },
               child: const Text('MobileScanner without Controller'),
+            ),
+            ElevatedButton(
+              onPressed: () {
+                Navigator.of(context).push(
+                  MaterialPageRoute(
+                    builder: (context) => const BarcodeScannerWithZoom(),
+                  ),
+                );
+              },
+              child: const Text('MobileScanner with zoom slider'),
             ),
           ],
         ),