+++ ios/Classes/SwiftMobileScannerPlugin.swift
@@ -2,6 +2,7 @@ import Flutter
 import MLKitVision
 import MLKitBarcodeScanning
 import AVFoundation
+import UIKit
 
 public class SwiftMobileScannerPlugin: NSObject, FlutterPlugin {
     
@@ -10,19 +11,52 @@ public class SwiftMobileScannerPlugin: NSObject, FlutterPlugin {
     
     /// The handler sends all information via an event channel back to Flutter
     private let barcodeHandler: BarcodeHandler
+
+    static var scanWindow: [CGFloat]?
+    
+    private static func isBarcodeInScanWindow(barcode: Barcode, imageSize: CGSize) -> Bool {
+        let scanwindow = SwiftMobileScannerPlugin.scanWindow!
+        let barcodeminX = barcode.cornerPoints![0].cgPointValue.x
+        let barcodeminY = barcode.cornerPoints![1].cgPointValue.y
+        
+        let barcodewidth = barcode.cornerPoints![2].cgPointValue.x - barcodeminX
+        let barcodeheight = barcode.cornerPoints![3].cgPointValue.y - barcodeminY
+        let barcodeBox = CGRect(x: barcodeminX, y: barcodeminY, width: barcodewidth, height: barcodeheight)
+
+        
+        let minX = scanwindow[0] * imageSize.width
+        let minY = scanwindow[1] * imageSize.height
+
+        let width = (scanwindow[2] * imageSize.width)  - minX
+        let height = (scanwindow[3] * imageSize.height) - minY
+
+        let scaledWindow =  CGRect(x: minX, y: minY, width: width, height: height)
+        
+        return scaledWindow.contains(barcodeBox)
+    }
     
     init(barcodeHandler: BarcodeHandler, registry: FlutterTextureRegistry) {
         self.mobileScanner = MobileScanner(registry: registry, mobileScannerCallback: { barcodes, error, image in
             if barcodes != nil {
-                let barcodesMap = barcodes!.map { barcode in
-                    return barcode.data
+                let barcodesMap = barcodes!.compactMap { barcode in
+                    if (SwiftMobileScannerPlugin.scanWindow != nil) {
+                        if (SwiftMobileScannerPlugin.isBarcodeInScanWindow(barcode: barcode, imageSize: image.size)) {
+                            return barcode.data
+                        } else {
+                            return nil
+                        }
+                    } else {
+                        return barcode.data
+                    }
                 }
                 if (!barcodesMap.isEmpty) {
-                    barcodeHandler.publishEvent(["name": "barcode", "data": barcodesMap, "image": FlutterStandardTypedData(bytes: image.jpegData(compressionQuality: 0.8)!)])
+                    barcodeHandler.publishEvent(["name": "barcode", "data": barcodesMap, "image": FlutterStandardTypedData(bytes: image.jpegData(compressionQuality: 0.8)!), "width": image.size.width, "height": image.size.height])
                 }
             } else if (error != nil){
                 barcodeHandler.publishEvent(["name": "error", "data": error!.localizedDescription])
             }
+        }, torchModeChangeCallback: { torchState in
+            barcodeHandler.publishEvent(["name": "torchState", "data": torchState])
         })
         self.barcodeHandler = barcodeHandler
         super.init()
@@ -49,6 +83,10 @@ public class SwiftMobileScannerPlugin: NSObject, FlutterPlugin {
             toggleTorch(call, result)
         case "analyzeImage":
             analyzeImage(call, result)
+        case "setScale":
+            setScale(call, result)
+        case "updateScanWindow":
+            updateScanWindow(call, result)
         default:
             result(FlutterMethodNotImplemented)
         }
@@ -64,7 +102,7 @@ public class SwiftMobileScannerPlugin: NSObject, FlutterPlugin {
 
         let formatList = formats.map { format in return BarcodeFormat(rawValue: format)}
         var barcodeOptions: BarcodeScannerOptions? = nil
-        
+
          if (formatList.count != 0) {
              var barcodeFormats: BarcodeFormat = []
              for index in formats {
@@ -123,6 +161,55 @@ public class SwiftMobileScannerPlugin: NSObject, FlutterPlugin {
         result(nil)
     }
     
+    /// Toggles the zoomScale
+    private func setScale(_ call: FlutterMethodCall, _ result: @escaping FlutterResult) {
+        var scale = call.arguments as? CGFloat
+        if (scale == nil) {
+            result(FlutterError(code: "MobileScanner",
+                                              message: "You must provide a scale when calling setScale!",
+                                              details: nil))
+            return
+        }
+        do {
+            try mobileScanner.setScale(scale!)
+        } catch MobileScannerError.zoomWhenStopped {
+            result(FlutterError(code: "MobileScanner",
+                                message: "Called setScale() while stopped!",
+                                details: nil))
+        } catch MobileScannerError.zoomError(let error) {
+            result(FlutterError(code: "MobileScanner",
+                                message: "Error while zooming.",
+                                details: error))
+        } catch {
+            result(FlutterError(code: "MobileScanner",
+                                message: "Error while zooming.",
+                                details: nil))
+        }
+        result(nil)
+    }
+    
+    /// Toggles the torch
+    func updateScanWindow(_ call: FlutterMethodCall, _ result: @escaping FlutterResult) {
+        let scanWindowData: Array? = (call.arguments as? [String: Any])?["rect"] as? [CGFloat]
+        SwiftMobileScannerPlugin.scanWindow = scanWindowData
+
+        result(nil)
+    }
+    
+    static func arrayToRect(scanWindowData: [CGFloat]?) -> CGRect? {
+        if (scanWindowData == nil) {
+            return nil
+        }
+
+        let minX = scanWindowData![0]
+        let minY = scanWindowData![1]
+
+        let width = scanWindowData![2]  - minX
+        let height = scanWindowData![3] - minY
+
+        return CGRect(x: minX, y: minY, width: width, height: height)
+    }
+    
     /// Analyzes a single image
     private func analyzeImage(_ call: FlutterMethodCall, _ result: @escaping FlutterResult) {
         let uiImage = UIImage(contentsOfFile: call.arguments as? String ?? "")
@@ -144,17 +231,5 @@ public class SwiftMobileScannerPlugin: NSObject, FlutterPlugin {
             }
         })
         result(nil)
-    }
-    
-    /// Observer for torch state
-    public override func observeValue(forKeyPath keyPath: String?, of object: Any?, change: [NSKeyValueChangeKey : Any]?, context: UnsafeMutableRawPointer?) {
-        switch keyPath {
-        case "torchMode":
-            // off = 0; on = 1; auto = 2;
-            let state = change?[.newKey] as? Int
-            barcodeHandler.publishEvent(["name": "torchState", "data": state])
-        default:
-            break
-        }
     }
 }