+++ lib/src/mobile_scanner_controller.dart
@@ -99,19 +99,21 @@ class MobileScannerController {
 
   bool isStarting = false;
 
-  bool? _hasTorch;
+  /// A notifier that provides availability of the Torch (Flash)
+  final ValueNotifier<bool?> hasTorchState = ValueNotifier(false);
 
   /// Returns whether the device has a torch.
   ///
   /// Throws an error if the controller is not initialized.
   bool get hasTorch {
-    if (_hasTorch == null) {
+    final hasTorch = hasTorchState.value;
+    if (hasTorch == null) {
       throw const MobileScannerException(
         errorCode: MobileScannerErrorCode.controllerUninitialized,
       );
     }
 
-    return _hasTorch!;
+    return hasTorch;
   }
 
   /// Set the starting arguments for the camera
@@ -124,11 +126,20 @@ class MobileScannerController {
     arguments['speed'] = detectionSpeed.index;
     arguments['timeout'] = detectionTimeoutMs;
 
+    /*    if (scanWindow != null) {
+      arguments['scanWindow'] = [
+        scanWindow!.left,
+        scanWindow!.top,
+        scanWindow!.right,
+        scanWindow!.bottom,
+      ];
+    } */
+
     if (formats != null) {
-      if (Platform.isAndroid) {
-        arguments['formats'] = formats!.map((e) => e.index).toList();
-      } else if (Platform.isIOS || Platform.isMacOS) {
+      if (kIsWeb || Platform.isIOS || Platform.isMacOS) {
         arguments['formats'] = formats!.map((e) => e.rawValue).toList();
+      } else if (Platform.isAndroid) {
+        arguments['formats'] = formats!.map((e) => e.index).toList();
       }
     }
     arguments['returnImage'] = true;
@@ -221,8 +232,9 @@ class MobileScannerController {
       );
     }
 
-    _hasTorch = startResult['torchable'] as bool? ?? false;
-    if (_hasTorch! && torchEnabled) {
+    final hasTorch = startResult['torchable'] as bool? ?? false;
+    hasTorchState.value = hasTorch;
+    if (hasTorch && torchEnabled) {
       torchState.value = TorchState.on;
     }
 
@@ -234,7 +246,7 @@ class MobileScannerController {
               startResult['videoHeight'] as double? ?? 0,
             )
           : toSize(startResult['size'] as Map? ?? {}),
-      hasTorch: _hasTorch!,
+      hasTorch: hasTorch,
       textureId: kIsWeb ? null : startResult['textureId'] as int?,
       webId: kIsWeb ? startResult['ViewID'] as String? : null,
     );
@@ -255,7 +267,7 @@ class MobileScannerController {
   ///
   /// Throws if the controller was not initialized.
   Future<void> toggleTorch() async {
-    final hasTorch = _hasTorch;
+    final hasTorch = hasTorchState.value;
 
     if (hasTorch == null) {
       throw const MobileScannerException(
@@ -294,6 +306,22 @@ class MobileScannerController {
         .then<bool>((bool? value) => value ?? false);
   }
 
+  /// Set the zoomScale of the camera.
+  ///
+  /// [zoomScale] must be within 0.0 and 1.0, where 1.0 is the max zoom, and 0.0
+  /// is zoomed out.
+  Future<void> setZoomScale(double zoomScale) async {
+    if (zoomScale < 0 || zoomScale > 1) {
+      throw const MobileScannerException(
+        errorCode: MobileScannerErrorCode.genericError,
+        errorDetails: MobileScannerErrorDetails(
+          message: 'The zoomScale must be between 0 and 1.',
+        ),
+      );
+    }
+    await _methodChannel.invokeMethod('setScale', zoomScale);
+  }
+
   /// Disposes the MobileScannerController and closes all listeners.
   ///
   /// If you call this, you cannot use this controller object anymore.
@@ -325,6 +353,8 @@ class MobileScannerController {
           BarcodeCapture(
             barcodes: parsed,
             image: event['image'] as Uint8List?,
+            width: event['width'] as double?,
+            height: event['height'] as double?,
           ),
         );
         break;
@@ -344,10 +374,12 @@ class MobileScannerController {
         _barcodesController.add(
           BarcodeCapture(
             barcodes: [
-              Barcode(
-                rawValue: barcode?['rawValue'] as String?,
-                rawBytes: barcode?['rawBytes'] as Uint8List?,
-              )
+              if (barcode != null)
+                Barcode(
+                  rawValue: barcode['rawValue'] as String?,
+                  rawBytes: barcode['rawBytes'] as Uint8List?,
+                  format: toFormat(barcode['format'] as int),
+                ),
             ],
           ),
         );
@@ -360,5 +392,11 @@ class MobileScannerController {
       default:
         throw UnimplementedError(name as String?);
     }
+  }
+
+  /// updates the native scanwindow
+  Future<void> updateScanWindow(Rect window) async {
+    final data = [window.left, window.top, window.right, window.bottom];
+    await _methodChannel.invokeMethod('updateScanWindow', {'rect': data});
   }
 }